refactor(core): move inline parsing out of domain
